move send invoice method from registrationService to emailService
